+++ tests/FactoryTest.php
@@ -73,6 +73,7 @@ class FactoryTest extends PHPUnit_Framework_TestCase
             array('codeship', 'CI_NAME', 'codeship'),
             array('wercker', 'WERCKER_MAIN_PIPELINE_STARTED', '1399372237'),
             array('shippable', 'SHIPPABLE', 'true'),
+            array('appveyor', 'APPVEYOR', 'True'),
             array('none', 'NONE', 'nonexisting-environment'),
         );
     }